+++ shark-browser.js
@@ -21,9 +21,10 @@ Object.assign(sharkUploadFile, {
 })
 
 /*
- * Expose Shark clients
+ * Expose Shark API
  */
 const Shark = require('./src/shark')
+const { isArray, isFunction, isObject, isString } = require('./src/utils/typecheck')
 
 Shark.Client = require('./src/clients/base-client')
 Shark.ActivityClient = require('./src/clients/activity-client')
@@ -33,9 +34,15 @@ Shark.ConsentClient = require('./src/clients/consent-client')
 Shark.ConsentChangesClient = require('./src/clients/consent-changes-client')
 Shark.ContactClient = require('./src/clients/contact-client')
 Shark.DescriptionClient = require('./src/clients/description-client')
+Shark.GroupClient = require('./src/clients/group-client')
 Shark.NotificationsClient = require('./src/clients/notifications-client')
 Shark.ServiceTokenClient = require('./src/service-token/browser')
 Shark.SubscriptionClient = require('./src/clients/subscription-client')
 Shark.UserClient = require('./src/clients/user-client')
+Shark.fetch = require('./src/utils/simple-fetch').simpleFetch
+Shark.isArray = isArray
+Shark.isFunction = isFunction
+Shark.isObject = isObject
+Shark.isString = isString
 
 module.exports = Shark